+++ lab5/lib/CMakeLists.txt
@@ -9,10 +9,10 @@ FetchContent_Declare(
 FetchContent_MakeAvailable(stb)
 
 if (ASSEMBLER)
-    file(GLOB SOURCES asm/invert_image.s)
+    file(GLOB SOURCES process_image.c asm/invert_image.s)
     set_source_files_properties(${SOURCES} PROPERTIES LANGUAGE ASM_NASM COMPILE_FLAGS "-g -F dwarf")
 else ()
-    file(GLOB SOURCES c/invert_image.c)
+    file(GLOB SOURCES process_image.c c/invert_image.c)
 endif ()
 
 add_library(lib STATIC ${SOURCES})
